requirements.txt
apache-libcloud
timestring
humanize
requests
shade
ruamel.yaml
